powerpc/vsx: implement vec_xl for vector_double
