+++ crates/core_arch/src/powerpc/altivec.rs
@@ -728,40 +728,8 @@ pub(crate) mod sealed {
         unsafe fn vec_xl(self, a: isize) -> Self::Result;
     }
 
-    macro_rules! impl_vec_xl {
-        ($fun:ident $notpwr9:ident / $pwr9:ident $ty:ident) => {
-            #[inline]
-            #[target_feature(enable = "altivec")]
-            #[cfg_attr(
-                all(test, not(target_feature = "power9-altivec")),
-                assert_instr($notpwr9)
-            )]
-            #[cfg_attr(all(test, target_feature = "power9-altivec"), assert_instr($pwr9))]
-            pub unsafe fn $fun(a: isize, b: *const $ty) -> t_t_l!($ty) {
-                let addr = (b as *const u8).offset(a);
-
-                let mut r = mem::MaybeUninit::uninit();
-
-                crate::ptr::copy_nonoverlapping(
-                    addr,
-                    r.as_mut_ptr() as *mut u8,
-                    mem::size_of::<t_t_l!($ty)>(),
-                );
-
-                r.assume_init()
-            }
-
-            #[unstable(feature = "stdarch_powerpc", issue = "111145")]
-            impl VectorXl for *const $ty {
-                type Result = t_t_l!($ty);
-                #[inline]
-                #[target_feature(enable = "altivec")]
-                unsafe fn vec_xl(self, a: isize) -> Self::Result {
-                    $fun(a, self)
-                }
-            }
-        };
-    }
+    // Use the impl_vec_xl macro from macros module
+    use crate::core_arch::powerpc::macros::impl_vec_xl;
 
     impl_vec_xl! { vec_xl_i8 lxvd2x / lxv i8 }
     impl_vec_xl! { vec_xl_u8 lxvd2x / lxv u8 }
